requirements.txt
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.0
kiwisolver==1.4.8
lxml==6.0.0
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
seaborn==0.13.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
xlsxwriter==3.2.5
